# k-vet-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "k-vet API",
    "description": "Veterinary practice management — single user, self-hosted.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/customer-emails/{id}": {
      "delete": {
        "tags": [
          "customers"
        ],
        "operationId": "deleteCustomerEmail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      },
      "patch": {
        "tags": [
          "customers"
        ],
        "operationId": "patchCustomerEmail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCustomerEmail"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "422": {
            "description": ""
          }
        }
      }
    },
    "/api/customers": {
      "get": {
        "tags": [
          "customers"
        ],
        "operationId": "listCustomers",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Text filter (name, number, …).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "archived",
            "in": "query",
            "description": "Include archived records (default: false).",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "customers"
        ],
        "operationId": "createCustomer",
        "responses": {
          "200": {
            "description": "An empty draft, ready for auto-save",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      }
    },
    "/api/customers/{id}": {
      "get": {
        "tags": [
          "customers"
        ],
        "operationId": "getCustomer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      },
      "patch": {
        "tags": [
          "customers"
        ],
        "operationId": "patchCustomer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCustomer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "422": {
            "description": ""
          }
        }
      }
    },
    "/api/customers/{id}/archive": {
      "post": {
        "tags": [
          "customers"
        ],
        "operationId": "archiveCustomer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      }
    },
    "/api/customers/{id}/emails": {
      "post": {
        "tags": [
          "customers"
        ],
        "operationId": "addCustomerEmail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerEmail"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "422": {
            "description": ""
          }
        }
      }
    },
    "/api/customers/{id}/unarchive": {
      "post": {
        "tags": [
          "customers"
        ],
        "operationId": "unarchiveCustomer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateCustomerEmail": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "email_type": {
            "$ref": "#/components/schemas/EmailType"
          }
        }
      },
...
      "Customer": {
        "type": "object",
        "required": [
          "id",
          "archived",
          "draft",
          "missing_fields",
          "has_invoice_address",
          "has_second_name",
          "emails",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "salutation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Salutation"
              }
            ]
          },
          "first_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "second_salutation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Salutation",
                "description": "Optional second person of the household, addressed on the invoice as well."
              }
            ]
          },
          "second_first_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "second_last_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "home_addon": {
            "type": [
              "string",
              "null"
            ]
          },
          "home_street": {
            "type": [
              "string",
              "null"
            ]
          },
          "home_zip": {
            "type": [
              "string",
              "null"
            ]
          },
          "home_city": {
            "type": [
              "string",
              "null"
            ]
          },
          "home_country": {
            "type": [
              "string",
              "null"
            ],
            "description": "ISO 3166-1 alpha-2; null falls back to `[invoice] default_country`."
          },
          "invoice_salutation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Salutation"
              }
            ]
          },
          "invoice_first_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "invoice_last_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "invoice_addon": {
            "type": [
              "string",
              "null"
            ]
          },
          "invoice_street": {
            "type": [
              "string",
              "null"
            ]
          },
          "invoice_zip": {
            "type": [
              "string",
              "null"
            ]
          },
          "invoice_city": {
            "type": [
              "string",
              "null"
            ]
          },
          "invoice_country": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ],
            "description": "Stored E.164 form, e.g. `+493012345678`."
          },
          "phone_display": {
            "type": [
              "string",
              "null"
            ],
            "description": "The same number in the familiar national format, e.g. `030 12345678`."
          },
          "warning_remark": {
            "type": [
              "string",
              "null"
            ]
          },
          "archived": {
            "type": "boolean"
          },
          "draft": {
            "type": "boolean"
          },
          "missing_fields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "has_invoice_address": {
            "type": "boolean",
            "description": "`true` when the invoice address group is complete and replaces the home address."
          },
          "has_second_name": {
            "type": "boolean"
          },
          "emails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerEmail"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CustomerEmail": {
        "type": "object",
        "required": [
          "id",
          "email",
          "email_type"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "email_type": {
            "$ref": "#/components/schemas/EmailType"
          }
        }
      },
...
      "EmailType": {
        "type": "string",
        "description": "Type of a customer email address.",
        "enum": [
          "private",
          "work",
          "other"
        ]
      },
...
      "PatchCustomerEmail": {
        "type": "object",
        "properties": {
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "email_type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EmailType"
              }
            ]
          }
        }
      },
...
      "Salutation": {
        "type": "string",
        "description": "Form of address; `familie` addresses a household.",
        "enum": [
          "frau",
          "herr",
          "familie"
        ]
      },
...
    }
  },
...
}